+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## Unreleased
 ### Added
 - Added mapping to and from ZAP alert ids and PTK module names
+- Example alerts
 
 ## 0.2.0
 ### Added